requirements.txt
openai>=1.30.0
requests>=2.32.0
python-dotenv>=1.0.1
Pillow>=10.4.0
